backend/requirements.txt
flask
python-dotenv
SQLAlchemy
psycopg2-binary
alembic
requests
oracledb